Added simple test for RapidApp::Responder::UserError
I added an ErrorGen controller to the ChinookDemo, though it
might have been more appropriate to make a new app for this
purpose.

This test only checks that the magic header and required
response data are present.  The more important test will be
when I get Selenium integrated, to test the actual ajax
receiving the message.

refs #108